Upcoming Brigade Horizon Bannur Road Bangalore
Reputed real estate organization Brigade Group has unveiled its new high-rise residential venture named Brigade Horizon in the City of Bangalore. This luxurious project is constructed on Mysore Road and has elegant homes and excellent connectivity to the development. With a unique conceptualization of having a healthy living space, this housing complex is nestled with lush green ambiance and world-class conveniences.

The integrated township of Brigade Horizon Bengaluru offers luxurious 2 BHK & 3 BHK apartments facilitated with modern conveniences. Italian marble floor for the living room, covered wooden floor for the master bedroom, and vitrified tiles for the bedroom enhance the elegance of the flats. Besides, every room is furnished with an anodized aluminum door frame and window fitting, which allows entry of sunlight and cross aeration.

Furthermore, there is a split AC provision in the living and master bedroom and window AC for the room. Anti-skid tiles have been placed on the bathroom floor along with a well-equipped imported fitting, shower cabin with a standing bath. The kitchen is set up in a contemporary style with modular cabinets, granite countertop, etc. 

This extensive master layout property in Bangalore has a gated appearance. The visual of this township is the paved pathway that leads to a lush green landscape having seating arrangements and a play area for kids. One attractive feature here is the Club which is a lavish clubhouse featuring a swimming pool, exercise room, yoga deck, etc. A shopping entrance to the Brigade Horizon Mysore Road further catches the attention of the home purchasers. Other amenities include an elevator, car parking, security guards, power backup, and many more.

For More Details:
Visit Here: https://bit.ly/3cEDmfw